Update oasis with ppx_core dependency.
The '-predicates' flag is necessary due to the issue here:
https://github.com/janestreet/ppx_core/issues/1

While the recommended fix is to edit the local _tags file, that
introduces a hard dependency on OCaml >= 4.02 due to some changes
in the ocamlbuild _tags syntax. To maintain compatibility with older
OCaml versions, specify the -predicates build flag as done here.